Move single building in its own view file
