+++ app/api_alpha/urls.py
@@ -4,6 +4,7 @@ from django.urls import re_path
 from rest_framework import routers
 
 from api_alpha.endpoints.buildings.list_create_buildings import ListCreateBuildings
+from api_alpha.endpoints.buildings.single_building import SingleBuilding
 from api_alpha.views import ActivateUser
 from api_alpha.views import AdsTokenView
 from api_alpha.views import ADSVectorTileView
@@ -30,7 +31,6 @@ from api_alpha.views import OrganizationView
 from api_alpha.views import PlotsVectorTileView
 from api_alpha.views import RequestPasswordReset
 from api_alpha.views import RNBAuthToken
-from api_alpha.views import SingleBuilding
 from api_alpha.views import SplitBuildings
 
 # Routers provide an easy way of automatically determining the URL conf.
